Update GETTING_STARTED.md
Fix the typo in the naming convention (Kpts with GCN) and complemented the list of models. The naming convention was changed after MICCAI submission (so the final model evaluation for the paper with multiple frames was trained on the GCN models). I mixed up the names afterwards in this Readme. I hope this does not cause confusion.